Backend/tripper/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==39.0.0
Deprecated==1.2.13
Django==4.1.5
django-background-tasks==1.2.5
django-compat==1.0.15
django-environ==0.9.0
frozenlist==1.3.3
gcloud==0.18.3
googleapis-common-protos==1.58.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
jwcrypto==1.4.2
multidict==6.0.4
oauth2client==4.1.3
openai==0.26.1
outcome==1.2.0
Pillow==9.4.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
pycryptodomex==3.16.0
pyparsing==3.0.9
Pyrebase4==4.6.0
PySocks==1.7.1
python-jwt==4.0.0
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
urllib3==1.26.14
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
